feat: 기존 페이지 shared url로 변경
